+++ app/src/main/java/com/example/agogolev/moneytracker/MainActivity.java
@@ -1,6 +1,5 @@
 package com.example.agogolev.moneytracker;
 
-
 import android.support.design.widget.NavigationView;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -11,13 +10,10 @@ import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.MenuItem;
 
-import java.util.ArrayList;
-import java.util.List;
 
 public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
 
@@ -26,7 +22,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     private ActionBarDrawerToggle toggle;
     private DrawerLayout drawerLayout;
     private NavigationView navigationView;
-    private RecyclerView recyclerView;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -34,7 +30,6 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         setContentView(R.layout.activity_main);
         setActionBar();
         setupDrawerLayout();
-        //initRecycleView();
         if (savedInstanceState == null){
             replaceFragment(new FragmentExpenses());
         }
@@ -61,12 +56,6 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         setTitle(getString(R.string.app_name));
     }
 
-    private void initRecycleView(){
-//        recyclerView = (RecyclerView) findViewById(R.id.list_of_expenses);
-//        recyclerView.setLayoutManager(new LinearLayoutManager(this));
-//        ExpensesAdapter expencesAdapter = new ExpensesAdapter(getExpenses());
-//        recyclerView.setAdapter(expencesAdapter);
-    }
 
     private void replaceFragment(Fragment fragment){
 
@@ -83,18 +72,6 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         }
     }
 
-    private List<Expense> getExpenses(){
-        List<Expense> expences = new ArrayList<>();
-        expences.add(new Expense("Cinema", "120"));
-        expences.add(new Expense("Cafe", "150"));
-        expences.add(new Expense("Car", "1200000"));
-        expences.add(new Expense("cake", "15"));
-        expences.add(new Expense("egg", "56"));
-        expences.add(new Expense("Tea", "65"));
-        expences.add(new Expense("Coffee Pele", "135"));
-        expences.add(new Expense("bread", "21"));
-        return expences;
-    }
 
     @Override
     public boolean onNavigationItemSelected(MenuItem item) {
@@ -125,11 +102,15 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 
     @Override
     public void onBackPressed() {
+        FragmentManager manager = getSupportFragmentManager();
 
         if (drawerLayout.isDrawerOpen(GravityCompat.START)) {
             drawerLayout.closeDrawer(GravityCompat.START);
-        } else {
+        } else if (manager.getBackStackEntryCount() == 1) {
+            supportFinishAfterTransition();
+        } else{
             super.onBackPressed();
         }
+
     }
 }
