+++ app/src/main/res/layout/fragment_expenses.xml
@@ -2,12 +2,12 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center">
+    android:layout_height="match_parent">
 
-    <TextView
-        android:layout_width="wrap_content"
+    <android.support.v7.widget.RecyclerView
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="@string/fragm_expenses"
-        android:textSize="@dimen/text_size_22"/>
+        android:id="@+id/list_of_expenses">
+    </android.support.v7.widget.RecyclerView>
+
 </LinearLayout>
